A tool that places a widget takes the decimals the board promises
The instructions every session reads say a widget "may take decimals, so
x=4.5 and w=3.25 are ordinary". Twelve of the seventeen placing tools typed
x/y/w/h as int, so a model that believed them got back "Input should be a
valid integer, got a number with a fractional part". It happened on the live
board yesterday: add_note at x=6.56 was refused.

The service, the schemas and the other five tools were fractional all along;
only these signatures were not. So this is the promise being kept, not a new
capability, and nothing that is drawn changes.

The test reads the catalogue rather than a list written by hand: every tool
that exposes x, y, w or h must type it as number. The thirteenth placing tool
is covered the day someone writes it, which is the failure mode that produced
this bug in the first place.

+++ backend/hud_mcp/charts.py
@@ -44,10 +44,10 @@ def register(server: MCPServer) -> None:
         colors: list[str] | None = None,
         unfilled: str | None = None,
         thresholds: list[dict] | None = None,
-        x: int | None = None,
-        y: int | None = None,
-        w: int | None = None,
-        h: int | None = None,
+        x: float | None = None,
+        y: float | None = None,
+        w: float | None = None,
+        h: float | None = None,
         description: str | None = None,
     ) -> str:
         """Draw a chart from data you supply inline.
